Private link error handling
