+++ common/environment/azure.go
@@ -1,0 +1,33 @@
+package environment
+
+
+type azureEnvironment struct {
+	Name                      string `json:"name"`
+	ServiceManagementEndpoint string `json:"serviceManagementEndpoint"`
+	ResourceManagerEndpoint   string `json:"resourceManagerEndpoint"`
+	ActiveDirectoryEndpoint   string `json:"activeDirectoryEndpoint"`
+}
+
+// based on github.com/Azure/go-autorest/autorest/azure/azureEnvironments.go
+var (
+	AzurePublicCloud = azureEnvironment{
+		Name:                      "PUBLIC",
+		ServiceManagementEndpoint: "https://management.core.windows.net/",
+		ResourceManagerEndpoint:   "https://management.azure.com/",
+		ActiveDirectoryEndpoint:   "https://login.microsoftonline.com/",
+	}
+
+	AzureUsGovernmentCloud = azureEnvironment{
+		Name:                      "USGOVERNMENT",
+		ServiceManagementEndpoint: "https://management.core.usgovcloudapi.net/",
+		ResourceManagerEndpoint:   "https://management.usgovcloudapi.net/",
+		ActiveDirectoryEndpoint:   "https://login.microsoftonline.us/",
+	}
+
+	AzureChinaCloud = azureEnvironment{
+		Name:                      "CHINA",
+		ServiceManagementEndpoint: "https://management.core.chinacloudapi.cn/",
+		ResourceManagerEndpoint:   "https://management.chinacloudapi.cn/",
+		ActiveDirectoryEndpoint:   "https://login.chinacloudapi.cn/",
+	}
+)
